Throw error when agent cannot be started
Previously if there's an error when starting an agent, the error is
effectively silenced. I'm guessing this is to suppress `already_started`
"errors". Unfortunately this also suppress any error

With this change, `already_started` will be ignored, but any other error
will blow up, making it easy to see what went wrong. I considered
logging an error and continuing, but it's simpler to blow up, and feels
aligned with "let it crash"